# node-api/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/get-all-banks": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-bank": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/create-bank": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "location": {
                  "type": "string",
                  "example": "Novi Sad"
                },
                "pib": {
                  "type": "string",
                  "example": "123456789"
                },
                "foundationYear": {
                  "type": "number",
                  "example": 1990
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/search-banks-older-than": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "year",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/search-banks-by-location": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "location",
            "in": "query",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}